requirements.txt
dacite==1.9.2
embreex==2.17.7.post6
geffnet==1.0.2
glob2==0.7
hydra-core==1.3.2
jaxtyping==0.3.0
matplotlib==3.10.1
mitsuba==3.6.4
omegaconf==2.3.0
opencv-python==4.11.0.86
protobuf==4.25
pyembree==0.1.12
pyexr==0.5.0
pyyaml==6.0.1
scipy==1.15.2
tensorboard==2.19.0
torch-tb-profiler==0.4.3
torchmetrics==1.7.0
torchmetrics[image]==1.7.0
tqdm==4.66.4
trimesh==3.23.5
